src/main/java/today/what_should_i_eat_today/domain/tag/application/TagService.java
package today.what_should_i_eat_today.domain.tag.application;

import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;
import today.what_should_i_eat_today.domain.tag.dao.TagRepository;
import today.what_should_i_eat_today.domain.tag.entity.Tag;

@Service
@RequiredArgsConstructor
@Transactional(readOnly = true)
public class TagService {
    private final TagRepository repository;

    public Page<Tag> getTagList(String name, Pageable pageable) {
        return repository.findByNameContains(name, pageable);
    }
}
